Made separate libraries
Made separate libraries for the different functions
