+++ home.h
@@ -1,0 +1,8 @@
+#ifndef HOME_H
+#define HOME_H
+
+void home(void);
+extern int bet, tokens;
+extern char response;
+
+#endif
