+++ race_funcs.h
@@ -1,0 +1,6 @@
+#ifndef RACE_FUNCS_H
+#define RACE_FUNCS_H
+
+void startRaceGamble(void);
+
+#endif
