+++ slots_funcs.h
@@ -1,0 +1,6 @@
+#ifndef SLOTS_FUNCS_H
+#define SLOTS_FUNCS_H
+
+void startSlotsGamble(void);
+
+#endif
